big refactor into different modules
