+++ .gitignore
@@ -1,0 +1,48 @@
+# Python specific
+*.pyc
+__pycache__/
+.Python/
+venv/
+.env/
+env/
+.pytest_cache/
+.mypy_cache/
+.pytype/
+.ruff_cache/
+
+# Streamlit specific
+.streamlit/cache/       # Streamlit's caching directory
+.streamlit/local_data/ # Local data created by Streamlit sessions (e.g., uploaded files)
+
+# IDE specific files
+.vscode/                # VS Code workspace files
+.idea/                  # IntelliJ/PyCharm IDEA files
+*.iml
+*.iws
+*.ipr
+
+# Operating System specific files
+.DS_Store               # macOS specific files
+Thumbs.db               # Windows thumbnail cache
+desktop.ini             # Windows system file
+ehthumbs.db
+.Trash-* # macOS trash files
+
+# Logs and temporary files
+*.log
+*.tmp
+*.temp
+*~                      # Backup files (e.g., from text editors)
+
+# Distribution and build artifacts
+dist/
+build/
+*.egg-info/
+.eggs/
+
+# Jupyter Notebook specific (if used)
+.ipynb_checkpoints/
+
+# Database files (if local, e.g., SQLite)
+*.db
+*.sqlite3
