+++ pages_ui/__init__.py
@@ -1,0 +1,5 @@
+from . import max_price
+from . import simulation
+from . import early_payments
+from . import monte_carlo_page
+from . import initial_costs
